horas/requirements-dev.txt
-r requirements.txt
black
invoke
mypy
pylint<2.8.0
pylint-django
